// src/types/enums.ts
export enum CohortType {
  MASTERING_BITCOIN = 'MASTERING_BITCOIN',
  LEARNING_BITCOIN_FROM_COMMAND_LINE = 'LEARNING_BITCOIN_FROM_COMMAND_LINE',
  PROGRAMMING_BITCOIN = 'PROGRAMMING_BITCOIN',
  BITCOIN_PROTOCOL_DEVELOPMENT = 'BITCOIN_PROTOCOL_DEVELOPMENT',
  MASTERING_LIGHTNING_NETWORK = 'MASTERING_LIGHTNING_NETWORK',
}
export enum UserRole {
  ADMIN = 'ADMIN',
  TEACHING_ASSISTANT = 'TEACHING_ASSISTANT',
  STUDENT = 'STUDENT',
}

export enum ComponentRating {
  NOT_AT_ALL = 'NOT_AT_ALL',
  SOMEWHAT = 'SOMEWHAT',
  HELPFUL = 'HELPFUL',
  VERY_HELPFUL = 'VERY_HELPFUL',
}

export enum CohortComponent {
  SESSION_INSTRUCTIONS = 'sessionInstructions',
  STUDY_MATERIAL = 'studyMaterial',
  GROUP_DISCUSSIONS = 'groupDiscussions',
  LOUNGE_DISCUSSIONS = 'loungeDiscussions',
  DEPUTY = 'deputy',
  TEACHING_ASSISTANTS = 'teachingAssistants',
  BITSHALA_CLUBS = 'bitshalaClubs',
  BITDEV_MEETUPS = 'bitdevMeetups',
  BITSPACE = 'bitspace',
  FELLOWSHIPS = 'fellowships',
}

export enum OpportunityInterest {
  DEVELOPER = 'DEVELOPER',
  DESIGNER_CREATIVE = 'DESIGNER_CREATIVE',
  EDUCATION_WRITING = 'EDUCATION_WRITING',
  PROGRAM_OPS = 'PROGRAM_OPS',
  LEGAL_ACCOUNTS = 'LEGAL_ACCOUNTS',
  BUILDING_STARTUP = 'BUILDING_STARTUP',
  HOSTING_CLUB = 'HOSTING_CLUB',
  JUST_STACKING = 'JUST_STACKING',
}

export enum FellowshipInterest {
  SILENT_PAYMENT_LIBRARY = 'SILENT_PAYMENT_LIBRARY',
  SILENT_PAYMENT_INDEXER = 'SILENT_PAYMENT_INDEXER',
  COINSELECTION = 'COINSELECTION',
  BITCOIN_CORE_REVIEW = 'BITCOIN_CORE_REVIEW',
  COINSWAP = 'COINSWAP',
  ANY_OTHER_PROJECT = 'ANY_OTHER_PROJECT',
}